feat: implement contact form API and refactor email sending logic
