--- src/app/api/send-email/route.js
@@ -1,54 +1,0 @@
-import nodemailer from "nodemailer";
-
-export async function POST(request) {
-  try {
-    const body = await request.json();
-    const { name, email, phone, subject, message } = body;
-
-    if (!name || !email || !message) {
-      return new Response(
-        JSON.stringify({ message: "Please fill in all required fields." }),
-        { status: 400 },
-      );
-    }
-    // Configure the transporter with custom SMTP settings
-    const transporter = nodemailer.createTransport({
-      host: process.env.SMTP_HOST, // Your SMTP host
-      port: process.env.SMTP_PORT, // SMTP port (usually 587 or 465 for secure)
-      secure: false,
-      auth: {
-        user: process.env.SMTP_USER, // SMTP username
-        pass: process.env.SMTP_PASS, // SMTP password
-      },
-    });
-
-    // Define email options
-    const mailOptions = {
-      from: process.env.SMTP_USER, // Sender details
-      to: process.env.ADMIN_EMAIL, // Admin's email
-      subject: subject || "No Subject",
-      text: `
-        Name: ${name}
-        Email: ${email}
-        Phone: ${phone || "Not provided"}
-        Message: ${message}
-      `,
-    };
-
-    // Send the email
-    await transporter.sendMail(mailOptions);
-
-    return new Response(
-      JSON.stringify({ message: "Email sent successfully!" }),
-      { status: 200 },
-    );
-  } catch (error) {
-    console.error("Error sending email:", error);
-    return new Response(
-      JSON.stringify({
-        message: "Error sending email. Please try again later.",
-      }),
-      { status: 500 },
-    );
-  }
-}
