+++ src/components/ContactSection.jsx
@@ -27,7 +27,7 @@ const ContactSection = () => {
     setIsLoading(true);
 
     try {
-      const response = await fetch("/api/send-email", {
+      const response = await fetch("/api/contact", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify(formData),
